requirements.txt
reportlab
groq
python-dotenv
streamlit